% Auto-generated: do not edit by hand
\name{agGrid}

\alias{agGrid}

\title{AgGrid component}

\description{
AgGrid is a component wrapped on ag-Grid. It takes two property, `columnDefs`, and `rowData` displays them.
}

\usage{
agGrid(id=NULL, columnDefs=NULL, gridApi=NULL, rowData=NULL,
selectedRows=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{columnDefs}{Unnamed list. columnDefs for ag-Grid.}

\item{gridApi}{Named list. gridApi for ag-Grid.}

\item{rowData}{Unnamed list. rowData for ag-Grid.}

\item{selectedRows}{Unnamed list. selectedRows for ag-Grid.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

