% Auto-generated: do not edit by hand
\name{apexChart}

\alias{apexChart}

\title{ApexChart component}

\description{
Wrapper for react-apexcharts library. For API and examples see:  https://github.com/apexcharts/react-apexcharts
}

\usage{
apexChart(id=NULL, height=NULL, options=NULL, series=NULL, type=NULL,
width=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{height}{Character | numeric. Possible values for height can be 100\%, 300px or 300 (by default is auto)}

\item{options}{Named list. The configuration object, see options on API (Reference)}

\item{series}{Unnamed list. The series is a set of data. To know more about the format of
the data, checkout Series docs on the website.}

\item{type}{Character. line, area, bar, pie, donut, scatter, bubble, heatmap, radialBar}

\item{width}{Character | numeric. Possible values for width can be 100\%, 400px or 400 (by default is 100\%)}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

