% Auto-generated: do not edit by hand
\name{renderJson}

\alias{renderJson}

\title{RenderJson component}

\description{

}

\usage{
renderJson(id=NULL, data=NULL, invert_theme=NULL, max_depth=NULL,
theme=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{data}{Named list. The value displayed in the input.}

\item{invert_theme}{Logical. }

\item{max_depth}{Numeric. Dash-assigned callback that should be called to report property changes
to Dash, to make them available for callbacks.}

\item{theme}{Named list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

