% Auto-generated: do not edit by hand
\name{s3Upload}

\alias{s3Upload}

\title{S3Upload component}

\description{

}

\usage{
s3Upload(id=NULL, label=NULL, value=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{label}{Character. A label that will be printed when this component is rendered.}

\item{value}{Character. The value displayed in the input.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

