% Auto-generated: do not edit by hand
\name{table}

\alias{table}

\title{Table component}

\description{

}

\usage{
table(columns=NULL, enableColumnSelection=NULL,
enableRangeSelection=NULL, enableRowSelection=NULL,
rows=NULL)
}

\arguments{
\item{columns}{Unnamed list. }

\item{enableColumnSelection}{Logical. }

\item{enableRangeSelection}{Logical. }

\item{enableRowSelection}{Logical. }

\item{rows}{Unnamed list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

